Moved out the behaviors
